third_party/blink/web_tests/platform/generic/fast/events/clipboard-clearData-expected.txt

Tests clipboard.clearData()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testDataTransfer.clearData(null) did not throw exception.
PASS testDataTransfer.getData("text") is "sample"
PASS testDataTransfer.getData("url") is "http://www.google.com/"
PASS testDataTransfer.getData("text/html") is "<em>Markup</em>"
PASS testDataTransfer.getData("custom-data") is "hello world"
PASS testDataTransfer.clearData("custom-data") did not throw exception.
PASS testDataTransfer.getData("text") is "sample"
PASS testDataTransfer.getData("url") is "http://www.google.com/"
PASS testDataTransfer.getData("text/html") is "<em>Markup</em>"
PASS testDataTransfer.getData("custom-data") is ""
PASS testDataTransfer.clearData() did not throw exception.
PASS testDataTransfer.getData("text") is ""
PASS testDataTransfer.getData("url") is ""
PASS testDataTransfer.getData("text/html") is ""
PASS testDataTransfer.getData("custom-data") is ""
PASS testDataTransfer.clearData(undefined) did not throw exception.
PASS testDataTransfer.getData("text") is ""
PASS testDataTransfer.getData("url") is ""
PASS testDataTransfer.getData("text/html") is ""
PASS testDataTransfer.getData("custom-data") is ""
PASS successfullyParsed is true

TEST COMPLETE
